WIP: Fusion Parser remove AbstractParser
also use xor as it reads clearer. (No functional diff as && is checked too.)
